requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.902
mypy-extensions==0.4.3
numpy==1.20.3
pathspec==0.8.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.8.3
pytz==2021.1
regex==2021.4.4
requests==2.25.1
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
atomicwrites==1.4.0
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.7
Babel==2.9.1
black==21.6b0
cattrs==1.5.0
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
constructs==10.0.5
cryptography==3.4.7
cssselect==1.1.0
distlib==0.3.2
Django==3.2.4
django-crispy-forms==1.12.0
djangorestframework==3.12.4
docutils==0.17.1
filelock==3.0.12
flake8==3.9.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
identify==2.2.10
idna==2.10
imagesize==1.2.0
incremental==21.3.0
iniconfig==1.1.1
isort==5.8.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==3.0.1
jmespath==0.10.0
jsii==1.30.0
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.902
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.20.3
packaging==20.9
parsel==1.6.0
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.13.0
priority==1.3.0
Protego==0.1.16
publication==0.0.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.8.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
w3lib==1.22.0
wrapt==1.12.1
zope.interface==5.4.0
